requirements.txt
tensorboardX
lmdb
tensorboard
torchvision
scikit-learn
scipy
tqdm
pytorch
numpy
transformers
datasets
tiktoken